pyproject.toml
[project]
name = "python-database"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Joao Victor", email = "jvcastelo.mec@gmail.com" }
]
dependencies = [
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "psycopg2>=2.9.10",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipykernel>=6.29.5",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/python_database"]
